Updated Makefile for cygwin/mingw build and fix misc warnings
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@145 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip/src/pjsip/sip_msg.c
@@ -1716,8 +1716,8 @@ PJ_DEF(pjsip_warning_hdr*) pjsip_warning_hdr_create(  pj_pool_t *pool,
 				     host->slen + 2 +	/* host */
 				     text->slen + 2);	/* text */
     hvalue.slen = pj_sprintf(hvalue.ptr, "%u %.*s \"%.*s\"",
-			     code, host->slen, host->ptr,
-			     text->slen, text->ptr);
+			     code, (int)host->slen, host->ptr,
+			     (int)text->slen, text->ptr);
 
     return pjsip_generic_string_hdr_create(pool, &str_warning, &hvalue);
 }
